ovl_en_trt Ok and Documented (#225)
* First pass: Matched everything but func_80A8E26C

* Matched func_80A8E26C

* Documented

* Cleanup

* Some renames

* Update EnTrt struct

* Move init chain out of the function

* Change unreferenced rodata implementation

* Texture stuff

* PR suggestions from other PRs that apply here

* Fix syntax error and other cleanup

* add cutscene state enum

* PR suggestions except color

* Update src/overlays/actors/ovl_En_Trt/z_en_trt.c


* Update src/overlays/actors/ovl_En_Trt/z_en_trt.c


* Update src/overlays/actors/ovl_En_Trt/z_en_trt.c


* Better bitflags

* Fix

+++ src/overlays/actors/ovl_Obj_Ghaka/z_obj_ghaka.c
@@ -97,7 +97,7 @@ void func_80B3C39C(ObjGhaka* this, GlobalContext* globalCtx) {
 }
 
 void func_80B3C4E0(ObjGhaka* this, GlobalContext* globalCtx) {
-    s32 talkState = func_80152498(&globalCtx->msgCtx);
+    u8 talkState = func_80152498(&globalCtx->msgCtx);
 
     if (talkState == 5) {
         if (func_80147624(globalCtx)) {
